授業中にこっそりHomeの中のformタグを別のクラスに分けました。
